refactor(archive): split pipeline internals into pipeline.rs
archive/mod.rs exceeded the 1000-line gate (1201 lines per the check).
Move plan_archive / query_buckets / commit_archive and their private
types (Parsed, Bucket, ArchivePlan, BucketWithResult) to a new
archive/pipeline.rs submodule.

mod.rs: 1200 → 915 lines (gate counts 916, well under 1000).
pipeline.rs: 312 lines.
Zero logic changes — pure structural move.
