+++ desktop/src-tauri/src/archive/mod.rs
@@ -15,9 +15,12 @@
 //! validation (sig/id + kind + p-tag + agent tag + frame=telemetry + author
 //! == agent) is applied fail-closed.
 
+mod pipeline;
 pub mod store;
 
-use nostr::{Event, JsonUtil};
+use pipeline::{commit_archive, plan_archive, query_buckets, BucketWithResult};
+
+use nostr::Event;
 use rusqlite::Connection;
 use serde::{Deserialize, Serialize};
 use tauri::State;
@@ -149,294 +152,6 @@ pub async fn archive_events(
         now,
         &conn,
     )
-}
-
-// ── Archive internals ────────────────────────────────────────────────────────
-
-/// A parsed, sig-verified candidate ready for further processing.
-struct Parsed {
-    event: Event,
-    raw_json: String,
-    matched_scope: MatchedScope,
-}
-
-/// One scope bucket: a set of candidates that share a scope type+value,
-/// with the relay filter already built and the subscription kinds loaded.
-struct Bucket {
-    scope_type_str: String,
-    scope_value: String,
-    allowed_kinds: Vec<u64>,
-    filter: serde_json::Value,
-    group: Vec<Parsed>,
-}
-
-/// Output of the sync planning phase.
-struct ArchivePlan {
-    buckets: Vec<Bucket>,
-    ephemeral: Vec<Parsed>,
-    /// Events already accounted as dropped during planning (no subscription,
-    /// unknown scope type, parse failure, bad sig).
-    pre_dropped: u32,
-}
-
-/// Phase 1 (sync): parse all candidates, group persistent ones into per-scope
-/// buckets, and load the subscription kinds for each bucket.
-///
-/// Returns an [`ArchivePlan`] with no `&Connection` remaining — safe to hold
-/// across `.await`.
-fn plan_archive(
-    candidates: Vec<ArchiveCandidate>,
-    identity_pk: &str,
-    relay_url: &str,
-    conn: &Connection,
-) -> Result<ArchivePlan, String> {
-    let mut persistent_raw: Vec<Parsed> = Vec::new();
-    let mut ephemeral: Vec<Parsed> = Vec::new();
-    let mut pre_dropped: u32 = 0;
-
-    for cand in candidates {
-        let event = match Event::from_json(&cand.raw_event_json) {
-            Ok(e) => e,
-            Err(_) => {
-                pre_dropped += 1;
-                continue;
-            }
-        };
-        if !event.verify_id() || !event.verify_signature() {
-            pre_dropped += 1;
-            continue;
-        }
-
-        if cand.matched_scope.scope_type.is_ephemeral() {
-            ephemeral.push(Parsed {
-                event,
-                raw_json: cand.raw_event_json,
-                matched_scope: cand.matched_scope,
-            });
-        } else {
-            persistent_raw.push(Parsed {
-                event,
-                raw_json: cand.raw_event_json,
-                matched_scope: cand.matched_scope,
-            });
-        }
-    }
-
-    // Group persistent candidates by (scope_type, scope_value).
-    use std::collections::HashMap;
-    let mut scope_groups: HashMap<(String, String), Vec<Parsed>> = HashMap::new();
-    for p in persistent_raw {
-        let key = (
-            p.matched_scope.scope_type.as_str().to_string(),
-            p.matched_scope.scope_value.clone(),
-        );
-        scope_groups.entry(key).or_default().push(p);
-    }
-
-    let mut buckets: Vec<Bucket> = Vec::with_capacity(scope_groups.len());
-    for ((scope_type_str, scope_value), mut group) in scope_groups {
-        // No subscription → drop the whole group.
-        let kinds_json = match store::get_subscription_kinds(
-            conn,
-            identity_pk,
-            relay_url,
-            &scope_type_str,
-            &scope_value,
-        )? {
-            Some(k) => k,
-            None => {
-                pre_dropped += group.len() as u32;
-                continue;
-            }
-        };
-
-        let allowed_kinds: Vec<u64> =
-            serde_json::from_str::<Vec<u64>>(&kinds_json).unwrap_or_default();
-
-        // Deduplicate by event id within the bucket.
-        let mut seen = std::collections::HashSet::new();
-        group.retain(|p| seen.insert(p.event.id.to_hex()));
-
-        let ids: Vec<String> = group.iter().map(|p| p.event.id.to_hex()).collect();
-
-        // Build a *scoped* relay filter: ids + scope tag + kinds.
-        let filter = match scope_type_str.as_str() {
-            "channel_h" => serde_json::json!({
-                "ids":   ids,
-                "#h":    [&scope_value],
-                "kinds": allowed_kinds,
-            }),
-            "referenced_e" => serde_json::json!({
-                "ids":   ids,
-                "#e":    [&scope_value],
-                "kinds": allowed_kinds,
-            }),
-            _ => {
-                pre_dropped += group.len() as u32;
-                continue;
-            }
-        };
-
-        buckets.push(Bucket {
-            scope_type_str,
-            scope_value,
-            allowed_kinds,
-            filter,
-            group,
-        });
-    }
-
-    Ok(ArchivePlan {
-        buckets,
-        ephemeral,
-        pre_dropped,
-    })
-}
-
-/// A bucket with the relay's response attached.
-struct BucketWithResult {
-    scope_type_str: String,
-    scope_value: String,
-    allowed_kinds: Vec<u64>,
-    group: Vec<Parsed>,
-    /// Event ids returned by the relay for the scoped filter.
-    returned_ids: std::collections::HashSet<String>,
-    /// True if the relay query failed (network error); entire group dropped.
-    relay_failed: bool,
-}
-
-/// Phase 2 (async): fire one relay query per bucket and collect results.
-///
-/// `state` is `&AppState` — a `Copy` reference — so no `!Send` value is held
-/// across `.await`.
-async fn query_buckets(buckets: Vec<Bucket>, state: &AppState) -> Vec<BucketWithResult> {
-    let mut results: Vec<BucketWithResult> = Vec::with_capacity(buckets.len());
-    for bucket in buckets {
-        let (returned_ids, relay_failed) = match query_relay(state, &[bucket.filter]).await {
-            Ok(evs) => (evs.iter().map(|e| e.id.to_hex()).collect(), false),
-            Err(_) => (std::collections::HashSet::new(), true),
-        };
-        results.push(BucketWithResult {
-            scope_type_str: bucket.scope_type_str,
-            scope_value: bucket.scope_value,
-            allowed_kinds: bucket.allowed_kinds,
-            group: bucket.group,
-            returned_ids,
-            relay_failed,
-        });
-    }
-    results
-}
-
-/// Phase 3 (sync): apply relay results and write accepted events to the store.
-fn commit_archive(
-    bucket_results: Vec<BucketWithResult>,
-    ephemeral: Vec<Parsed>,
-    pre_dropped: u32,
-    identity_pk: &str,
-    relay_url: &str,
-    now: i64,
-    conn: &Connection,
-) -> Result<ArchiveBatchResult, String> {
-    let mut persisted: u32 = 0;
-    let mut dropped: u32 = pre_dropped;
-
-    // ── Persistent path ──────────────────────────────────────────────────────
-    for result in bucket_results {
-        if result.relay_failed {
-            dropped += result.group.len() as u32;
-            continue;
-        }
-
-        for p in result.group {
-            let eid = p.event.id.to_hex();
-
-            // Relay proof: event was returned for the scoped filter.
-            if !result.returned_ids.contains(&eid) {
-                dropped += 1;
-                continue;
-            }
-
-            // Kind enforcement: event.kind must be in the subscription's list.
-            if !result
-                .allowed_kinds
-                .contains(&(p.event.kind.as_u16() as u64))
-            {
-                dropped += 1;
-                continue;
-            }
-
-            // The relay returning this event for {ids, #h/#e, kinds} IS the
-            // proof of scope membership. Use scope_value directly; no local
-            // tag re-derivation (which would incorrectly drop h-less events
-            // matched via the relay's StoredEvent.channel_id fallback).
-            store::upsert_archived_event(
-                conn,
-                identity_pk,
-                relay_url,
-                &eid,
-                p.event.kind.as_u16() as i64,
-                &p.event.pubkey.to_hex(),
-                p.event.created_at.as_secs() as i64,
-                &p.raw_json,
-                now,
-            )?;
-            store::upsert_event_scope(
-                conn,
-                identity_pk,
-                relay_url,
-                &eid,
-                &result.scope_type_str,
-                &result.scope_value,
-                now,
-            )?;
-            persisted += 1;
-        }
-    }
-
-    // ── Ephemeral path (owner_p) ─────────────────────────────────────────────
-    // Fully local validation — no relay query.
-    for p in ephemeral {
-        match validate_ephemeral_frame(
-            &p.event,
-            identity_pk,
-            &p.matched_scope.scope_value,
-            conn,
-            identity_pk,
-            relay_url,
-        ) {
-            Ok(()) => {}
-            Err(_) => {
-                dropped += 1;
-                continue;
-            }
-        }
-
-        let eid = p.event.id.to_hex();
-        store::upsert_archived_event(
-            conn,
-            identity_pk,
-            relay_url,
-            &eid,
-            p.event.kind.as_u16() as i64,
-            &p.event.pubkey.to_hex(),
-            p.event.created_at.as_secs() as i64,
-            &p.raw_json,
-            now,
-        )?;
-        store::upsert_event_scope(
-            conn,
-            identity_pk,
-            relay_url,
-            &eid,
-            "owner_p",
-            &p.matched_scope.scope_value,
-            now,
-        )?;
-        persisted += 1;
-    }
-
-    Ok(ArchiveBatchResult { persisted, dropped })
 }
 
 /// Validate an ephemeral observer frame (kind 24200) against ALL local rules.
